Factored out UncalibratedEventTimesMove into its own file.
